Fix deadlock when shutting down JobExecutor.
Previously, it was possible that at the beginning of shutdown the queue
was not empty, but became empty at the exit of the critical section,
which would leads to an infinite wait()'ing later on for the event that
all queued jobs are finished (it never comes because queue is empty
already and does not accept new jobs).

To resolve this: add a terminating flag so that no new jobs would
be accepted while the executor is still running, and change the locking
method for waiting for the queue to empty.